scripts/redact_cli_py/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

azure-core==1.17.0
azure-storage-blob==12.8.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4; python_version >= '3'
cryptography==3.4.7; python_version >= '3.6'
dacite==1.6.0
idna==3.2; python_version >= '3'
isodate==0.6.0
jsonpointer==2.1
msrest==0.6.21
oauthlib==3.1.1; python_version >= '3.6'
pillow==8.3.1
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pypdfium==0.0.15
requests-oauthlib==1.3.0
requests==2.26.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
shapely==1.7.1
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
